requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
APScheduler==3.10.4
playwright==1.44.0
openpyxl==3.1.2
Werkzeug==3.0.3